Prompt for confirmation when destroy is used

+++ internal/cmd/destroy.go
@@ -11,6 +11,8 @@ import (
 	"github.com/spf13/cobra"
 )
 
+var flagConfirmDestroy bool
+
 func newDestroyCommand(site *site.Site) *cobra.Command {
 
 	cmd := &cobra.Command{
@@ -24,22 +26,38 @@ func newDestroyCommand(site *site.Site) *cobra.Command {
 
 	commandsRequiringSite = append(commandsRequiringSite, cmd.Use)
 
+	cmd.Flags().BoolVar(&flagConfirmDestroy, "confirm-destroy", false, "Confirm destruction of your site (doesn't require a prompt).")
+
 	return cmd
 }
 
 func runDestroy(cmd *cobra.Command, args []string, site *site.Site) {
 
-	// Stop the WordPress site.
-	err := site.StopWordPress()
-	if err != nil {
-		console.Error(err, flagDebugMode)
+	var confirmDestroy = false
+
+	if flagConfirmDestroy {
+		confirmDestroy = true
+	} else {
+		confirmDestroy = console.PromptConfirm(fmt.Sprintf("Are you sure you want to destroy %s? %s", aurora.Bold(aurora.Blue(site.StaticConfig.SiteName)), aurora.Bold(aurora.Yellow("This operation is destructive and cannot be undone."))), false)
 	}
 
-	// Remove the site's folder in the config directory.
-	err = os.RemoveAll(site.StaticConfig.SiteDirectory)
-	if err != nil {
-		console.Error(err, flagDebugMode)
+	if confirmDestroy {
+		// Stop the WordPress site.
+		err := site.StopWordPress()
+		if err != nil {
+			console.Error(err, flagDebugMode)
+		}
+
+		// Remove the site's folder in the config directory.
+		err = os.RemoveAll(site.StaticConfig.SiteDirectory)
+		if err != nil {
+			console.Error(err, flagDebugMode)
+		}
+
+		console.Success(fmt.Sprintf("Your site, %s, has been completely destroyed.", aurora.Bold(aurora.Blue(site.StaticConfig.SiteName))))
+		return
 	}
 
-	console.Success(fmt.Sprintf("Your site, %s, has been completely destroyed.", aurora.Bold(aurora.Blue(site.StaticConfig.SiteName))))
+	console.Error(fmt.Errorf("site destruction cancelled. No data has been lost"), flagDebugMode)
+
 }
